requirements.txt
yfinance
pandas
discord-webhook
python-dotenv